+++ examples/instance_template/additional_disks/main.tf
@@ -15,16 +15,16 @@
  */
 
 provider "google" {
-  credentials = "${file(var.credentials_path)}"
-  project     = "${var.project_id}"
-  region      = "${var.region}"
-  version     = "~> 1.19"
+  credentials = file(var.credentials_path)
+  project     = var.project_id
+  region      = var.region
+  version     = "~> 2.7"
 }
 
 module "instance_template" {
   source          = "../../../modules/instance_template"
-  subnetwork      = "${var.subnetwork}"
-  service_account = "${var.service_account}"
+  subnetwork      = var.subnetwork
+  service_account = var.service_account
   name_prefix     = "additional-disks"
 
   additional_disks = [
@@ -48,3 +48,4 @@ module "instance_template" {
     },
   ]
 }
+