+++ examples/umig/static_ips/variables.tf
@@ -36,11 +36,15 @@ variable "num_instances" {
 }
 
 variable "static_ips" {
-  type        = "list"
+  type        = list(string)
   description = "List of static IPs for VM instances"
 }
 
 variable "service_account" {
-  type        = "map"
-  description = "Service account email address and scopes"
+  default = null
+  type = object({
+    email  = string
+    scopes = set(string)
+  })
+  description = "Service account to attach to the instance. See https://www.terraform.io/docs/providers/google/r/compute_instance_template.html#service_account."
 }