+++ test/fixtures/umig/named_ports/main.tf
@@ -20,12 +20,12 @@ locals {
 
 module "umig_named_ports" {
   source           = "../../../../examples/umig/named_ports"
-  credentials_path = "${local.credentials_path}"
-  project_id       = "${var.project_id}"
+  credentials_path = local.credentials_path
+  project_id       = var.project_id
   region           = "us-central1"
-  subnetwork       = "${google_compute_subnetwork.main.name}"
+  subnetwork       = google_compute_subnetwork.main.name
   num_instances    = 4
-  service_account  = "${var.service_account}"
+  service_account  = var.service_account
 
   named_ports = [
     {
@@ -38,3 +38,4 @@ module "umig_named_ports" {
     },
   ]
 }
+