+++ CHANGELOG.md
@@ -7,6 +7,10 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Supported version of Terraform is 0.12. [#16]
+
 ## [0.2.0] - 2019-05-30
 
 ### Added
@@ -23,3 +27,4 @@ project adheres to [Semantic Versioning](http://semver.org/).
 [0.2.0]: https://github.com/terraform-google-modules/terraform-google-vm/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/terraform-google-modules/terraform-google-vm/releases/tag/v0.1.0
 [#10]: https://github.com/terraform-google-modules/terraform-google-vm/pull/10
+[#16]: https://github.com/terraform-google-modules/terraform-google-vm/pull/16