+++ CHANGELOG.md
@@ -7,6 +7,12 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.0.0] - 2019-07-31
+
+### Added
+
+- `preemptible` variable on the Instance Template submodule. [#14]
+
 ### Changed
 
 - Supported version of Terraform is 0.12. [#16]
@@ -23,8 +29,10 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - `instance_template`, `mig`, and `umig` modules
 - examples, basic tests and fixtures for all modules
 
-[Unreleased]: https://github.com/terraform-google-modules/terraform-google-vm/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/terraform-google-modules/terraform-google-vm/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/terraform-google-modules/terraform-google-vm/compare/v0.2.0...v1.0.0
 [0.2.0]: https://github.com/terraform-google-modules/terraform-google-vm/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/terraform-google-modules/terraform-google-vm/releases/tag/v0.1.0
 [#10]: https://github.com/terraform-google-modules/terraform-google-vm/pull/10
+[#14]: https://github.com/terraform-google-modules/terraform-google-vm/pull/14
 [#16]: https://github.com/terraform-google-modules/terraform-google-vm/pull/16